fix(order): retype quantity to be a float

+++ src/book_management/mod.rs
@@ -1,14 +1,12 @@
 pub mod traded_instruments;
 
-use std::sync::Weak;
-
 use traded_instruments::Instrument;
 
 use crate::exchange_connectivity::ExchangeType;
 
 pub trait Order {
-    fn new(instrument: Instrument, exchange: ExchangeType, quantity: usize, price: f64) -> Self;
-    fn quantity(&self) -> usize;
+    fn new(instrument: Instrument, exchange: ExchangeType, quantity: f64, price: f64) -> Self;
+    fn quantity(&self) -> f64;
     fn price(&self) -> f64;
     fn instrument(&self) -> Instrument;
     fn exchange(&self) -> ExchangeType;
@@ -18,12 +16,12 @@ pub trait Order {
 pub struct Bid {
     instrument: Instrument,
     exchange: ExchangeType,
-    quantity: usize,
+    quantity: f64,
     price: f64,
 }
 
 impl Order for Bid {
-    fn new(instrument: Instrument, exchange: ExchangeType, quantity: usize, price: f64) -> Self {
+    fn new(instrument: Instrument, exchange: ExchangeType, quantity: f64, price: f64) -> Self {
         Bid {
             instrument: instrument,
             exchange: exchange,
@@ -32,7 +30,7 @@ impl Order for Bid {
         }
     }
 
-    fn quantity(&self) -> usize {
+    fn quantity(&self) -> f64 {
         self.quantity
     }
 
@@ -51,14 +49,14 @@ impl Order for Bid {
 
 #[derive(Debug)]
 pub struct Ask {
-    quantity: usize,
+    quantity: f64,
     price: f64,
     instrument: Instrument,
     exchange: ExchangeType,
 }
 
 impl Order for Ask {
-    fn new(instrument: Instrument, exchange: ExchangeType, quantity: usize, price: f64) -> Self {
+    fn new(instrument: Instrument, exchange: ExchangeType, quantity: f64, price: f64) -> Self {
         Ask {
             instrument: instrument,
             exchange: exchange,
@@ -67,7 +65,7 @@ impl Order for Ask {
         }
     }
 
-    fn quantity(&self) -> usize {
+    fn quantity(&self) -> f64 {
         self.quantity
     }
